moved ModEMConfig
