+++ mtpy/modeling/modem.py
@@ -33,10 +33,9 @@ import mtpy.analysis.pt as mtpt
 import mtpy.imaging.mtcolors as mtcl
 import mtpy.imaging.mtplottools as mtplottools
 import mtpy.modeling.ws3dinv as ws
-import mtpy.utils.configfile as mtcfg
 import mtpy.utils.exceptions as mtex
 # vtk tools
-from mtpy.modeling.ModEM import ModEMError, Data, Model, Covariance
+from mtpy.modeling.ModEM import Data, Model
 
 try:
     from evtk.hl import gridToVTK, pointsToVTK
@@ -66,59 +65,6 @@ except ImportError:
 # ==============================================================================
 # Write inversion parameters to a config type file
 # ==============================================================================
-class ModEM_Config(object):
-    """
-    read and write configuration files for how each inversion is run
-    """
-
-    def __init__(self, **kwargs):
-        self.cfg_dict = {'ModEM_Inversion_Parameters': {}}
-
-        for key in kwargs.keys():
-            setattr(self, key, kwargs[key])
-
-    def write_config_file(self, save_dir=None,
-                          config_fn_basename='ModEM_inv.cfg'):
-        """
-        write a config file based on provided information
-        """
-
-        if save_dir is None:
-            save_dir = os.getcwd()
-
-        cfg_fn = os.path.join(save_dir, config_fn_basename)
-
-        if self.cfg_dict is not None:
-            mtcfg.write_dict_to_configfile(self.cfg_dict,
-                                           cfg_fn)
-
-    def add_dict(self, fn=None, obj=None):
-        """
-        add dictionary based on file name or object
-        """
-
-        if fn is not None:
-            if fn.endswith('.rho'):
-                m_obj = Model()
-                m_obj.read_model_file(fn)
-            elif fn.endswith('.dat'):
-                m_obj = Data()
-                m_obj.read_data_file(fn)
-            elif fn.endswith('.cov'):
-                m_obj = Covariance()
-                m_obj.read_cov_fn(fn)
-        elif obj is not None:
-            m_obj = obj
-
-        else:
-            raise ModEMError('Need to input a file name or object')
-
-        add_dict = m_obj.get_parameters()
-
-        for key in add_dict.keys():
-            self.cfg_dict['ModEM_Inversion_Parameters'][key] = add_dict[key]
-
-
 # ==============================================================================
 # Manipulate the model to test structures or create a starting model
 # ==============================================================================
